worker/requirements.txt
aio_pika==9.4.1
boto3==1.34.81
pandas==2.2.2
pyarrow==15.0.2
transformers==4.39.3
Pillow==10.3.0
sentencepiece==0.1.99
